requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
einops==0.8.1
eval_type_backport==0.2.2
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.0.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.37.1
six==1.17.0
smmap==5.0.2
sympy==1.13.3
threadpoolctl==3.6.0
timm==1.0.19
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.67.1
triton==2.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
